MyFirstPackage

Documentation for MyFirstPackage.

MyFirstPackage.AbstractLBConfig

MyFirstPackage.D2Q9

MyFirstPackage.curl

MyFirstPackage.curl

MyFirstPackage.equilibrium_density

MyFirstPackage.momentum

MyFirstPackage.step!

MyFirstPackage.AbstractLBConfig — Type

AbstractLBConfig{D, N}

An abstract type for lattice Boltzmann configurations. D is the dimension of the lattice. N is the number of velocities.

MyFirstPackage.D2Q9 — Type

D2Q9 <: AbstractLBConfig{2, 9}

A lattice Boltzmann configuration for 2D, 9-velocity model.

MyFirstPackage.curl — Method

curl(u::AbstractMatrix{Point2D{T}})

Compute the curl of the momentum field in 2D, which is defined as:

\[∂u_y/∂x−∂u_x/∂y\]

MyFirstPackage.curl — Method

curl(u::AbstractMatrix{Point3D{T}})

Compute the curl of the momentum field in 3D, which is defined as:

\[∂u_y/∂x−∂u_x/∂y, ∂u_y/∂x−∂u_x/∂y, ∂u_y/∂x−∂u_x/∂y, \]

MyFirstPackage.equilibrium_density — Method

equilibrium_density(lb::AbstractLBConfig, ρ, u)

Compute the equilibrium density of the fluid from the total density and the momentum.

MyFirstPackage.momentum — Method

momentum(lb::AbstractLBConfig, rho::Cell)

Compute the momentum of the fluid from the density of the fluid.

MyFirstPackage.step! — Method

step!(lb::LatticeBoltzmann)

Perform a single step of the lattice Boltzmann simulation.